pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.36.0",
    "google-generativeai>=0.8.5",
    "requests>=2.32.5",
    "sift-stack-py>=0.8.5",
    "streamlit>=1.49.1",
]
